+++ astrocut/footprint_cutout.py
@@ -11,7 +11,7 @@ from astropy.table import Table, Column
 from cachetools import TTLCache, cached
 from spherical_geometry.polygon import SphericalPolygon
 
-from .Cutout import Cutout
+from .cutout import Cutout
 
 FFI_TTLCACHE = TTLCache(maxsize=10, ttl=900)  # Cache for FFI footprint files
 
@@ -255,3 +255,36 @@ def get_ffis(s3_footprint_cache: str) -> Table:
     ffis = Table(ffis)
 
     return ffis
+
+
+def ra_dec_crossmatch(all_ffis: Table, coordinates: SkyCoord, cutout_size, arcsec_per_px: int = 21) -> Table:
+    """
+    Returns the Full Frame Images (FFIs) whose footprints overlap with a cutout of a given position and size.
+
+    Parameters
+    ----------
+    all_ffis : `~astropy.table.Table`
+        Table of FFIs to crossmatch with the cutout.
+    coordinates : str or `astropy.coordinates.SkyCoord` object
+        The position around which to cutout.
+        It may be specified as a string ("ra dec" in degrees)
+        or as the appropriate `~astropy.coordinates.SkyCoord` object.
+    cutout_size : int, array-like, `~astropy.units.Quantity`
+        The size of the cutout array. If ``cutout_size``
+        is a scalar number or a scalar `~astropy.units.Quantity`,
+        then a square cutout of ``cutout_size`` will be created.  If
+        ``cutout_size`` has two elements, they should be in ``(ny, nx)``
+        order.  Scalar numbers in ``cutout_size`` are assumed to be in
+        units of pixels. `~astropy.units.Quantity` objects must be in pixel or
+        angular units.
+    arcsec_per_px : int, optional
+        Default 21. The number of arcseconds per pixel in an image. Used to determine
+        the footprint of the cutout. Default is the number of arcseconds per pixel in
+        a TESS image.
+
+    Returns
+    -------
+    matching_ffis : `~astropy.table.Table`
+        Table containing information about FFIs whose footprints overlap those of the cutout.
+    """
+    return FootprintCutout.ra_dec_crossmatch(all_ffis, coordinates, cutout_size, arcsec_per_px)
