+++ astrocut/image_cutout.py
@@ -14,7 +14,7 @@ from s3path import S3Path
 
 from . import log
 from .exceptions import DataWarning, InputWarning, InvalidInputError
-from .Cutout import Cutout
+from .cutout import Cutout
 
 
 class ImageCutout(Cutout, ABC):
@@ -398,3 +398,40 @@ class ImageCutout(Cutout, ABC):
         np.subtract(255, norm_img, out=norm_img, where=invert)
 
         return norm_img
+    
+
+def normalize_img(img_arr: np.ndarray, stretch: str = 'asinh', minmax_percent: Optional[List[int]] = None, 
+                  minmax_value: Optional[List[int]] = None, invert: bool = False) -> np.ndarray:
+    """
+    Apply given stretch and scaling to an image array.
+
+    Parameters
+    ----------
+    img_arr : array
+        The input image array.
+    stretch : str
+        Optional, default 'asinh'. The stretch to apply to the image array.
+        Valid values are: asinh, sinh, sqrt, log, linear
+    minmax_percent : array
+        Optional. Interval based on a keeping a specified fraction of pixels (can be asymmetric) 
+        when scaling the image. The format is [lower percentile, upper percentile], where pixel
+        values below the lower percentile and above the upper percentile are clipped.
+        Only one of minmax_percent and minmax_value shoul be specified.
+    minmax_value : array
+        Optional. Interval based on user-specified pixel values when scaling the image.
+        The format is [min value, max value], where pixel values below the min value and above
+        the max value are clipped.
+        Only one of minmax_percent and minmax_value should be specified.
+    invert : bool
+        Optional, default False.  If True the image is inverted (light pixels become dark and vice versa).
+
+    Returns
+    -------
+    response : array
+        The normalized image array, in the form in an integer arrays with values in the range 0-255.
+    """
+    return ImageCutout.normalize_img(img_arr=img_arr,
+                                     stretch=stretch,
+                                     minmax_percent=minmax_percent,
+                                     minmax_value=minmax_value,
+                                     invert=invert)
