+++ astrocut/tests/test_cube_cut.py
@@ -12,10 +12,10 @@ from astropy.io import fits
 from astropy.time import Time
 from astropy.wcs import FITSFixedWarning
 
-from ..cube_cut import CutoutFactory, cube_cut
+from ..cutout_factory import CutoutFactory, cube_cut
 from ..exceptions import InputWarning
-from ..CubeFactory import CubeFactory
-from ..TicaCubeFactory import TicaCubeFactory
+from ..cube_factory import CubeFactory
+from ..tica_cube_factory import TicaCubeFactory
 from .utils_for_test import create_test_ffis
